sistema54-app_nol_sched_print/backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
pydantic==2.6.1
pydantic-settings==2.1.0
jinja2==3.1.3
weasyprint>=67.0
python-multipart==0.0.9
email-validator==2.1.0.post1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
fpdf2==2.7.6
apscheduler==3.10.4
pyotp==2.9.0
qrcode[pil]==7.4.2
pillow==10.2.0